backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Brotli==1.1.0
cffi==1.16.0
chardet==5.2.0
cssselect2==0.7.0
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
fonttools==4.49.0
graphviz==0.20.1
html5lib==1.1
packaging==24.1
pillow==10.2.0
pip-review==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
pydotplus==2.0.2
pydyf==0.9.0
PyJWT==1.7.1
pyparsing==3.1.1
pyphen==0.14.0
python-dotenv==1.0.1
pytz==2023.3.post1
reportlab==4.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.9.0
weasyprint==61.1
webencodings==0.5.1
yattag==1.15.2
zopfli==0.2.3
